Add option to spoof API version
