+++ app/src/main/java/tk/wasdennnoch/androidn_ify/XposedHook.java
@@ -141,6 +141,7 @@ public class XposedHook implements IXposedHookLoadPackage, IXposedHookZygoteInit
         // Has to be hooked in every app as every app creates own instances of the Notification.Builder
         NotificationHooks.hook(lpparam.classLoader);
 
+
         // This actually is only used in the system process, but every app has access, so just to be sure hook everything
         if (ConfigUtils.qs().enable_qs_editor) {
             try {
@@ -149,6 +150,11 @@ public class XposedHook implements IXposedHookLoadPackage, IXposedHookZygoteInit
                 XposedBridge.hookAllMethods(classCMStatusBarManager, "publishTileAsUser", XC_MethodReplacement.DO_NOTHING);
             } catch (Throwable ignore) {
             }
+        }
+
+        ConfigUtils.notifications().loadSpoofAPIApps();
+        if (ConfigUtils.notifications().spoofAPIApps.contains(lpparam.packageName)) {
+            XposedHelpers.setStaticIntField(Build.VERSION.class, "SDK_INT", 24);
         }
 
     }
