+++ app/src/main/java/tk/wasdennnoch/androidn_ify/ui/SpoofAPIActivity.java
@@ -1,0 +1,8 @@
+package tk.wasdennnoch.androidn_ify.ui;
+
+public class SpoofAPIActivity extends AppListActivity {
+    @Override
+    protected String getKey() {
+        return "notification_spoof_api_version";
+    }
+}
