+++ app/src/main/res/layout/activity_app_list.xml
