+++ app/src/main/res/values/strings.xml
@@ -356,5 +356,8 @@
     <string name="battery_detail_switch_summary">Reduces performance and background data</string>
     <string name="qs_add_notes_title">Notes</string>
     <string name="qs_add_notes">This is only used for apps that provide broadcast tiles (such as \"Custom Quick Settings\"). Please read the instructions in the tile provider app before using this.</string>
+    <string name="notification_spoof_api_apps_title">Spoof API 24</string>
+    <string name="notification_spoof_api_version_title">Spoof API 24</string>
+    <string name="notification_spoof_api_version">Spoof these apps into thinking they\'re running on API 24</string>
 
 </resources>
